Rename functions for more intuitive pattern
