+++ signet-program/test-utils/constants.ts
@@ -7,3 +7,8 @@ export const SECP256K1_CURVE_ORDER = BigInt(
 export const ANCHOR_EMIT_CPI_CALL_BACK_DISCRIMINATOR = Buffer.from([
   0xe4, 0x45, 0xa5, 0x2e, 0x51, 0xcb, 0x9a, 0x1d,
 ]);
+
+export const eventNames = {
+  signatureRequested: "signatureRequestedEvent",
+  signatureResponded: "signatureRespondedEvent",
+} as const;
